requirements.txt
flask
requests
gpiozero
python-dotenv
RPi.GPIO
